Fully working rollbacks.
I had to manually implement transactions. Maybe this can be done better.
The main problem was that I use foreign keys in many tables and a row
needs to already be committed to the database so it can be referred to
via a foreign key. Once committed SQLAlchemy does not treat it as part
of the current transaction anymore. I also have no idea how to test
 this as I hope to never having to rollback by careful checks of the
 data.

Also waveform and station information uploaders now appear to fully
work.