Added extensive station uploading test.
Tests the created files and database entries. Furthermore the filepath
table now has a row "is_managed_by_seishub" to also be able to index
files that the user might not want SeisHub to mess with.
Also changed the URL of the station resource uploading to
"/event_based_data/resource/station". If all resource can be accessed in
this manner it should make the API simpler.